refactor: styled 파일 이동
