+++ components/StartPage.styled.ts
